requirements/prod.txt
django-filter==25.2
inflection==0.5.1
python-dotenv==1.2.2
aiofiles==24.1.0
aiogram==3.15.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.10.0
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
distlib==0.3.9
distro==1.9.0
Django==5.2.7
django-cors-headers==4.7.0
django-unfold==0.67.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.29.0
drf-yasg==1.21.11
filelock==3.18.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
ioutil==1.0.3
jiter==0.7.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
local==1.0.0
magic-filter==1.0.12
multidict==6.1.0
openai==1.55.1
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.3
pillow==11.2.1
platformdirs==4.3.6
propcache==0.2.0
pyarrow==22.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pyperclip==1.11.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.3
rpds-py==0.30.0
selenium==4.26.1
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
speedtest-cli==2.1.3
sqlparse==0.5.3
srutil==1.0.3
tabulate==0.9.0
toml==0.10.2
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.2.3
virtualenv==20.29.3
websocket-client==1.8.0
wikipedia==1.4.0
wsproto==1.2.0
yarl==1.18.0
python-decouple==3.8